Searching examples...logn complexity median question remaining

+++ src/main/java/sqrtofaNum.java
@@ -4,14 +4,14 @@ public class sqrtofaNum {
     }
 
     public static int sqRoot(int x){
-        int low=1,high=x;
+        long low=1,high=x;
         if(x==0||x==1)
             return x;
 
-        int ans=0;
+        long ans=0;
         while (low<=high){
-            int mid=(low+high)/2;
-            if(mid*mid == x) return mid;
+            long mid=(low+high)/2;
+            if(mid*mid == x) return (int)mid;
             if(mid*mid<x){//since we are returning floor it might not be perfect sqrt in that case lower bound is answer
                 low=mid+1;
                 ans=mid;
@@ -19,6 +19,6 @@ public class sqrtofaNum {
                 high=mid-1;
             }
         }
-        return ans;
+        return (int)ans;
     }
 }
